Enable yank/paste from vim and tmux to system clipboard
I utilized https://evertpot.com/osx-tmux-vim-copy-paste-clipboard/ to
get this working on OSX.

I had to tweak the original for this line:

`set-option -g default-command "reattach-to-user-namespace -l zsh"`

Originally it had bash. In the future I may want to determine what shell
is running, but for now this is fine because I'm using zsh across all my
machines.

This resolves #1.