Refresh tmux and zsh config
Made a switch to Windows for a hot minute and got everything setup with
some nice changes in wsl2 + windows terminal. Ended up getting a MacBook
Pro at work and had to set it up all again. This seems to work
seemlessly and thus needed to update the old dotfiles.